test(sandbox): make the sandbox suite a shared contract, like the store contract
Groundwork for a second sandbox implementation, and the answer to "keep the two paths
as similar as possible": the thing that keeps them honest is a shared SUITE, not a
shared description. Mirrors packages/db/test/store-contract.ts, which already does this
for the SQLite and Postgres drivers.

The 13 adversarial tests move into sandbox-contract.ts, parameterised by a factory, and
code-sandbox.test.ts becomes three lines that run them against the Node sandbox. A
Cloudflare implementation runs the same file, so anything visible to the code being run
— the tools surface, call_tool, console.log, what return does, the shape of a failure,
every limit — has to match or the suite fails.

The isolation MECHANICS are expected to differ and should: Node needs an in-context
prelude to close the constructor.constructor escape, while Workers disallows eval
outright and never had that hole.

Verified against Cloudflare's Worker Loader locally (wrangler dev, no deploy) before
writing any of it. Dynamic Workers went open beta in March 2026 for all paid Workers
plans — no waitlist, and Derive already qualifies via Containers/Queues/DO, so my
earlier "blocked on beta access" was simply out of date. The probe established:

  - `[[worker_loaders]] binding = "LOADER"` works in local dev
  - the isolate's env contains ONLY what is passed; env.LOADER is undefined, so the
    parent's bindings are not inherited and there is no empty-env trick to arrange
  - eval and the Function constructor throw "Code generation from strings disallowed"
  - globalOutbound must be a real Fetcher (a plain {fetch} object is rejected, and
    ctx.exports.X() needs an options argument)
  - globalOutbound: null blocks the internet outright — the default-deny egress the
    container substrate cannot express, since enableInternet there is one boolean
  - a runaway sync loop is NOT bounded by the runtime, so a host wall clock is required

The implementation itself is NOT committed. Two reasons, both worth stating: knip
correctly flags it as an unused file (nothing can import it until worker.ts and the
wrangler binding are wired), and its contract test cannot run in-process — a fake
loader executing the module on this event loop hangs on the runaway-loop case instead
of failing it, which is the same deadlock shape as calling execFileSync beside an
in-process server. Committing it would add exactly the parked, half-tested code this
branch has been arguing against. The draft is kept out of tree pending a wrangler-dev
harness that can host the real isolate.
